// src/server/model/iItem.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { IAttributeGroupLine } from './iAttributeGroupLine';
import { IChildItemDetail } from './iChildItemDetail';
import { IDocument } from './iDocument';
import { IItemStockPrice } from './iItemStockPrice';
import { IOpeningStockTx } from './iOpeningStockTx';

export interface IItem { 
    disabled?: boolean;
    longDescription?: string;
    manufacturerName?: string;
    enableBarcodePrinting?: boolean;
    showItemOnHomePage?: boolean;
    compatibleWith?: Array<string>;
    alternateUnitName?: string;
    tradeItemGroupName?: string;
    alternateUnitId?: number;
    openingStockTx?: IOpeningStockTx;
    childItemDetailLines?: Array<IChildItemDetail>;
    usePrimaryRate?: boolean;
    alternatePartNo?: Array<string>;
    tradeItemGroupId?: number;
    wholeSalePrice?: number;
    manufacturerId?: number;
    itemAttributeGroupId?: number;
    attributeGroupLine?: IAttributeGroupLine;
    stockPriceList?: Array<IItemStockPrice>;
    barcodeTracked?: boolean;
    alternateUnitMrp?: number;
    purchaseRtnAccountLedgerId?: number;
    sellingPriceDiscPercentage?: number;
    dealerPriceDiscPercentage?: number;
    purchaseRtnAccountLedgerName?: string;
    purchaseAccountLedgerId?: number;
    consumeChildItemsInSale?: boolean;
    itemAttributeGroupName?: string;
    purchaseAccountLedgerName?: string;
    alternateUnitRetailPrice?: number;
    alternateUnitLandingPrice?: number;
    landingPriceDisPercentage?: number;
    alternateUnitWholeSalePrice?: number;
    purchasePriceDisPercentage?: number;
    alternateUnitSellingPrice?: number;
    wholesalePriceDiscPercentage?: number;
    alternateUnitPurchasePrice?: number;
    useMrpForRatesCalculation?: boolean;
    itemgroupId?: number;
    productCode?: string;
    unitName?: string;
    unitId?: number;
    sellingPrice?: number;
    itemgroupName?: string;
    itemImages?: Array<IDocument>;
    maxStock?: number;
    openingRate?: number;
    maintainStock?: boolean;
    purchasePrice?: number;
    serviceId?: number;
    packingType?: string;
    minStock?: number;
    barCodeFile?: string;
    hsnCode?: string;
    dealerPrice?: number;
    rateCode?: string;
    landingPrice?: number;
    batchEnabled?: boolean;
    serviceName?: string;
    gstIncPrice?: number;
    openingStock?: number;
    masterItemId?: number;
    shelf?: string;
    mrp?: number;
    ledgerName?: string;
    ledgerId?: number;
    printName?: string;
    taxClassId?: number;
    taxClassName?: string;
    rtnLedgerName?: string;
    rtnLedgerId?: number;
    name?: string;
    description?: string;
    modificationdate?: Date;
    impCompanyGuid?: string;
    productUserName?: string;
    productUserId?: number;
    logInfo?: string;
    importedId?: number;
    creationdate?: Date;
    id?: number;
    jacksontype?: string;
}