// src/server/model/iLedgerGroup.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ILedgerGroup { 
    type?: ILedgerGroup.TypeEnum;
    system?: boolean;
    parentLedgerGroupName?: string;
    parentLedgerGroupId?: number;
    orderInBalanceSheet?: number;
    includeInPayableReceivable?: boolean;
    typeInt?: number;
    name?: string;
    description?: string;
    modificationdate?: Date;
    impCompanyGuid?: string;
    productUserName?: string;
    productUserId?: number;
    logInfo?: string;
    importedId?: number;
    creationdate?: Date;
    id?: number;
    jacksontype?: string;
}
export namespace ILedgerGroup {
    export type TypeEnum = 'LIABILITIES' | 'ASSETS' | 'EXPENSES' | 'INCOME' | 'PRIMARY';
    export const TypeEnum = {
        LIABILITIES: 'LIABILITIES' as TypeEnum,
        ASSETS: 'ASSETS' as TypeEnum,
        EXPENSES: 'EXPENSES' as TypeEnum,
        INCOME: 'INCOME' as TypeEnum,
        PRIMARY: 'PRIMARY' as TypeEnum
    };
}